+++ docs/source/conf.py
@@ -88,7 +88,7 @@ copyright = '2019, johneiser'
 author = 'johneiser'
 
 # The full version, including alpha/beta/rc tags
-release = '0.1.1'
+release = '0.1.2'
 
 
 # -- General configuration ---------------------------------------------------