+++ docs/source/changelog.rst
@@ -2,6 +2,17 @@
 Changelog
 =========
 
+0.1.2
+^^^^^
+
+April 12, 2019
+
+- Enabled interactive modules by restoring stdin to tty
+- Added modules: analyze/disassemble/x86, analyze/disassemble/x64
+- added images: tools/radare2
+- Changed image: tools/capstone (and thus modules: disassemble/*) to return only instructions, nothing else - leave the formatted disassembly to analyze/disassemble/*
+- Enabled test.py to handle errors gracefully
+
 0.1.1
 ^^^^^